Revert "homepage & fix bottomsheet"
