+++ lib/main.dart
@@ -1,12 +1,11 @@
 import 'dart:ui';
 
 import 'package:coba/screens/details_room.dart';
-import 'package:coba/screens/navigator.dart';
+import 'package:coba/screens/home_screen.dart';
 import 'package:coba/screens/login_screen.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:get/get.dart';
 
 Future<void> main() async {
   WidgetsFlutterBinding.ensureInitialized();
@@ -26,13 +25,13 @@ class MyApp extends StatelessWidget {
     SystemChrome.setSystemUIOverlayStyle(SystemUiOverlayStyle(
         systemNavigationBarColor: Colors.grey.shade400,
         statusBarIconBrightness: Brightness.light));
-    return GetMaterialApp(
+    return MaterialApp(
       title: 'mPurbaya',
       theme: ThemeData(
         primarySwatch: Colors.orange,
       ),
       debugShowCheckedModeBanner: false,
-      home: MainPage(),
+      home: const RoomDetailsPage(),
     );
   }
 }
