+++ lib/screens/home_screen.dart
@@ -1,317 +1,54 @@
-// ignore_for_file: prefer_const_constructors
+// ignore_for_file: prefer_const_constructors, sized_box_for_whitespace
 
-import 'package:carousel_slider/carousel_slider.dart';
-import 'package:coba/screens/details_room.dart';
+import 'package:coba/constant/color_constant.dart';
 import 'package:flutter/material.dart';
-import 'package:get/get.dart';
-import 'package:google_fonts/google_fonts.dart';
+import 'package:flutter_svg/flutter_svg.dart';
+import 'package:flutter_vector_icons/flutter_vector_icons.dart';
 
-final List<String> imgList = [
-  'assets/room2.jpg',
-  'assets/room2.jpg',
-  'assets/room2.jpg',
-  'assets/room2.jpg',
-];
-
-final List<Widget> imageSliders = imgList
-    .map((item) => Container(
-          child: Container(
-            child: ClipRRect(
-                borderRadius: BorderRadius.all(Radius.circular(15.0)),
-                child: Image.asset(item, fit: BoxFit.cover, width: 1000.0)),
-          ),
-        ))
-    .toList();
-
-int currentCarousel = 0;
-
-final CarouselController carouselController = CarouselController();
-
-class HomeScreen extends StatefulWidget {
-  HomeScreen({Key? key}) : super(key: key);
+class MainPage extends StatefulWidget {
+  const MainPage({Key? key}) : super(key: key);
 
   @override
-  _HomeScreenState createState() => _HomeScreenState();
+  _MainPageState createState() => _MainPageState();
 }
 
-class _HomeScreenState extends State<HomeScreen> {
+class _MainPageState extends State<MainPage> {
+  int currentIndex = 0;
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      body: SingleChildScrollView(
-        physics: BouncingScrollPhysics(),
-        child: Column(
-          children: [_header(), _carousel(), _body()],
-        ),
-      ),
-    );
-  }
-
-  Widget _header() {
-    return Padding(
-      padding: const EdgeInsets.fromLTRB(20, 30, 20, 0),
-      child: Row(
-        children: [
-          Column(
-            mainAxisAlignment: MainAxisAlignment.start,
-            crossAxisAlignment: CrossAxisAlignment.start,
-            // ignore: prefer_const_literals_to_create_immutables
-            children: [
-              Text("Selamat Datang",
-                  style: GoogleFonts.poppins(
-                      fontSize: 14, fontWeight: FontWeight.w400)),
-              Text("Vindy Bagus Prasetyo",
-                  style: GoogleFonts.poppins(
-                      color: Colors.orange,
-                      fontSize: 14,
-                      fontWeight: FontWeight.w700)),
-            ],
-          ),
-          Spacer(),
-          CircleAvatar(
-            backgroundColor: Colors.orange,
-            radius: 25,
-            child: Text("P",
-                style: GoogleFonts.poppins(
-                  color: Colors.white,
-                )),
-          )
+      bottomNavigationBar: BottomNavigationBar(
+        elevation: 9,
+        currentIndex: currentIndex,
+        type: BottomNavigationBarType.fixed,
+        onTap: (newValue) {
+          setState(() {
+            currentIndex = newValue;
+          });
+        },
+        items: [
+          _bottomIcon("assets/home-fill.svg", "assets/home.svg"),
+          _bottomIcon("assets/shop-bag-fill.svg", "assets/shop-bag.svg"),
+          _bottomIcon("assets/profile.svg", "assets/profile.svg"),
         ],
       ),
     );
   }
+}
 
-  Widget _carousel() {
-    return Padding(
-      padding: const EdgeInsets.fromLTRB(20, 25, 20, 10),
-      child: Column(
-        children: [
-          ClipRRect(
-            borderRadius: BorderRadius.circular(15),
-            child: Container(
-              child: CarouselSlider(
-                  carouselController: carouselController,
-                  options: CarouselOptions(
-                      height: Get.height * 0.2,
-                      autoPlay: true,
-                      viewportFraction: 1.0,
-                      enlargeCenterPage: false,
-                      autoPlayInterval: Duration(seconds: 7),
-                      autoPlayAnimationDuration: Duration(milliseconds: 800),
-                      onPageChanged: (index, reason) {
-                        setState(() {
-                          currentCarousel = index;
-                        });
-                      }),
-                  items: imageSliders),
-            ),
-          ),
-          Row(
-            mainAxisAlignment: MainAxisAlignment.center,
-            children: imgList.asMap().entries.map((entry) {
-              return GestureDetector(
-                onTap: () => carouselController.animateToPage(entry.key),
-                child: Container(
-                  width: 8.0,
-                  height: 8.0,
-                  margin: EdgeInsets.symmetric(vertical: 8.0, horizontal: 4.0),
-                  decoration: BoxDecoration(
-                      shape: BoxShape.circle,
-                      color: currentCarousel == entry.key
-                          ? Colors.orange
-                          : Colors.grey.shade400),
-                ),
-              );
-            }).toList(),
-          )
-        ],
-      ),
-    );
-  }
-
-  Widget _body() {
-    return Column(
-      mainAxisAlignment: MainAxisAlignment.start,
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        Padding(
-          padding: EdgeInsets.symmetric(horizontal: 20),
-          child: Text(
-              "Cari penawaran terbaik yang berkualitas pada Hotel Purbaya ...",
-              style: GoogleFonts.poppins(
-                  fontSize: 14, fontWeight: FontWeight.w400)),
-        ),
-        _tipeKamar(),
-        _wisata(),
-      ],
-    );
-  }
-
-  Widget _tipeKamar() {
-    return Column(
-      mainAxisAlignment: MainAxisAlignment.start,
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        Padding(
-          padding: const EdgeInsets.symmetric(horizontal: 20, vertical: 10),
-          child: Text("Tipe Kamar",
-              style: GoogleFonts.poppins(
-                  fontSize: 18, fontWeight: FontWeight.w600)),
-        ),
-        Container(
-            height: Get.height * 0.2,
-            child: ListView.builder(
-                padding: EdgeInsets.symmetric(horizontal: 20),
-                physics: BouncingScrollPhysics(),
-                scrollDirection: Axis.horizontal,
-                itemCount: imgList.length,
-                itemBuilder: (context, index) {
-                  return typeRoomCard(imgList, index);
-                })),
-      ],
-    );
-  }
-
-  Widget typeRoomCard(List product, index) {
-    return GestureDetector(
-      onTap: () {
-        Get.to(() => RoomDetailsPage());
-        // print("dadada");
-      },
-      child: ClipRRect(
-        borderRadius: BorderRadius.circular(15),
-        child: Card(
-          elevation: 1,
-          shape: RoundedRectangleBorder(
-            borderRadius: BorderRadius.circular(15.0),
-          ),
-          color: Colors.white,
-          child: Padding(
-            padding: const EdgeInsets.all(10.0),
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                Expanded(
-                  child: ClipRRect(
-                    borderRadius: BorderRadius.circular(15),
-                    child: Container(
-                      width: Get.width * 0.45,
-                      child: Image.asset(
-                        product[index].toString(),
-                        fit: BoxFit.cover,
-                      ),
-                    ),
-                  ),
-                ),
-                Padding(
-                  padding: const EdgeInsets.only(left: 5, top: 5),
-                  child: Column(
-                    crossAxisAlignment: CrossAxisAlignment.start,
-                    children: [
-                      Text(
-                        "Tipe A",
-                        style: GoogleFonts.poppins(
-                          fontSize: 18,
-                          fontWeight: FontWeight.w600,
-                        ),
-                      ),
-                      Row(
-                        children: [
-                          Text(
-                            "IDR 350.000 ",
-                            style: GoogleFonts.poppins(
-                                fontSize: 10, color: Colors.orange),
-                          ),
-                          Text(
-                            "/ Malam",
-                            style: GoogleFonts.poppins(
-                              fontSize: 10,
-                              fontWeight: FontWeight.w600,
-                            ),
-                          ),
-                        ],
-                      ),
-                    ],
-                  ),
-                )
-              ],
-            ),
-          ),
+_bottomIcon(String activeIcon, String icon) {
+  return BottomNavigationBarItem(
+      activeIcon: Container(
+        height: 20,
+        child: SvgPicture.asset(
+          activeIcon,
         ),
       ),
-    );
-  }
-
-  _wisata() {
-    return Column(
-      mainAxisAlignment: MainAxisAlignment.start,
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        Padding(
-          padding: const EdgeInsets.symmetric(horizontal: 20, vertical: 10),
-          child: Text("Wisata Terdekat",
-              style: GoogleFonts.poppins(
-                  fontSize: 18, fontWeight: FontWeight.w600)),
-        ),
-        Padding(
-          padding: const EdgeInsets.fromLTRB(0, 0, 0, 20),
-          child: Container(
-              height: Get.width * 0.45,
-              child: ListView.builder(
-                  padding: EdgeInsets.symmetric(horizontal: 20),
-                  physics: BouncingScrollPhysics(),
-                  scrollDirection: Axis.horizontal,
-                  itemCount: imgList.length,
-                  itemBuilder: (context, index) {
-                    return _listWisata(imgList, index);
-                  })),
-        ),
-      ],
-    );
-  }
-
-  Widget _listWisata(List product, index) {
-    return ClipRRect(
-      borderRadius: BorderRadius.circular(15),
-      child: Card(
-        elevation: 1,
-        shape: RoundedRectangleBorder(
-          borderRadius: BorderRadius.circular(15.0),
-        ),
-        color: Colors.white,
-        child: Padding(
-          padding: const EdgeInsets.all(10.0),
-          child: Column(
-            mainAxisAlignment: MainAxisAlignment.center,
-            crossAxisAlignment: CrossAxisAlignment.center,
-            children: [
-              Expanded(
-                child: ClipRRect(
-                  borderRadius: BorderRadius.circular(15),
-                  child: Container(
-                    width: Get.width * 0.35,
-                    height: 20,
-                    child: Image.asset(
-                      product[index].toString(),
-                      fit: BoxFit.cover,
-                    ),
-                  ),
-                ),
-              ),
-              Padding(
-                padding: const EdgeInsets.only(left: 5, top: 5),
-                child: Text(
-                  "Telaga Sarangan",
-                  style: GoogleFonts.poppins(
-                    fontSize: 12,
-                    fontWeight: FontWeight.w600,
-                  ),
-                ),
-              )
-            ],
-          ),
+      icon: Container(
+        height: 20,
+        child: SvgPicture.asset(
+          icon,
         ),
       ),
-    );
-  }
+      label: "");
 }
