--- lib/screens/profile_screen.dart
@@ -1,17 +1,0 @@
-import 'package:flutter/material.dart';
-
-class ProfileScreen extends StatefulWidget {
-  const ProfileScreen({ Key? key }) : super(key: key);
-
-  @override
-  _ProfileScreenState createState() => _ProfileScreenState();
-}
-
-class _ProfileScreenState extends State<ProfileScreen> {
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      child: Center(child: Text("Profile"),)
-    );
-  }
-}
