--- lib/screens/reservation_screen.dart
@@ -1,17 +1,0 @@
-import 'package:flutter/material.dart';
-
-class ReservationScreen extends StatefulWidget {
-  const ReservationScreen({ Key? key }) : super(key: key);
-
-  @override
-  _ReservationScreenState createState() => _ReservationScreenState();
-}
-
-class _ReservationScreenState extends State<ReservationScreen> {
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      child: Center(child: Text("Reservasi"),)
-    );
-  }
-}
