+++ pubspec.yaml
@@ -28,36 +28,38 @@ environment:
 # the latest version available on pub.dev. To see which dependencies have newer
 # versions available, run `flutter pub outdated`.
 dependencies:
-  carousel_slider: ^4.0.0
-  cloud_firestore: ^2.5.4
-  cupertino_icons: ^1.0.2
-  firebase_auth: ^3.1.4
-  firebase_core: ^1.8.0
   flutter:
     sdk: flutter
-  flutter_signin_button: ^2.0.0
+  splashscreen: ^1.2.0
+
+  # The following adds the Cupertino Icons font to your application.
+  # Use with the CupertinoIcons class for iOS style icons.
+  cupertino_icons: ^1.0.2
   flutter_svg: ^0.23.0+1
-  flutter_vector_icons: ^1.0.0
+  flutter_signin_button: ^2.0.0
+  cloud_firestore: ^2.5.4
+  firebase_auth: ^3.1.4
+  firebase_core: ^1.8.0
   fluttertoast: ^8.0.8
   get: ^4.1.4
-  google_fonts: ^2.1.0
-  google_sign_in: ^5.2.0
   material_design_icons_flutter: ^5.0.6295
+  flutter_vector_icons: ^1.0.0
+  google_sign_in: ^5.2.0
   provider: ^5.0.0
-  splashscreen: ^1.2.0
 dev_dependencies:
-  
+  flutter_test:
+    sdk: flutter
+
   # The "flutter_lints" package below contains a set of recommended lints to
   # encourage good coding practices. The lint set provided by the package is
   # activated in the `analysis_options.yaml` file located at the root of your
   # package. See that file for information about deactivating specific lint
   # rules and activating additional ones.
   flutter_lints: ^1.0.0
-  flutter_test:
-    sdk: flutter
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
+
 # The following section is specific to Flutter.
 flutter:
   # The following line ensures that the Material Icons font is
@@ -71,10 +73,13 @@ flutter:
     - assets/logo-putih.png
     - assets/google.png
     - assets/
+
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.dev/assets-and-images/#resolution-aware.
+
   # For details regarding adding assets from package dependencies, see
   # https://flutter.dev/assets-and-images/#from-packages
+
   # To add custom fonts to your application, add a fonts section here,
   # in this "flutter" section. Each entry in this list should have a
   # "family" key with the font family name, and a "fonts" key with a
